+++ CHANGELOG.md
@@ -14,6 +14,16 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) rules.
 * ```Fixed``` for any bug fixes.
 * ```Security``` in case of vulnerabilities.
 
+## [2021.1.3.10] - 2021-03-10
+
+### Added
+- complete loading alarm settings
+- complete editing alarm
+- complete saving new alarm
+
+### Changed
+- loading setting and data table immediately after change in selection
+
 ## [2021.1.3.9] - 2021-03-09
 
 ### Added